Changed message display from div to alert
